% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taxonomy.R
\name{check_taxonomy}
\alias{check_taxonomy}
\title{Check if a species translation table exists}
\usage{
check_taxonomy()
}
\value{
Logical (\code{TRUE} if a file exists, \code{FALSE} if not, with a message).
}
\description{
This function checks if at least one .csv file is present in \verb{inst/extdata/SpeciesTranslationTable/}.
This is required for use of the pfw_species and pfw_filter functions.
If no file exists, it prompts the user to download it using \code{update_taxonomy()}, or place a manually downloaded file there.
}
\keyword{internal}
