% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celebrate.R
\name{.celebrate}
\alias{.celebrate}
\title{Celebrate!}
\usage{
.celebrate(name = NULL)
}
\arguments{
\item{name}{(Optional) Your name, if you want!}
}
\value{
A celebration!
}
\description{
This function is just if you want to celebrate finishing some
particularly difficult code. Have a little fun with it--
I used it to celebrate finishing this R package!
}
\keyword{internal}
