% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_region_lookup.R
\name{load_region_lookup}
\alias{load_region_lookup}
\title{Load the region lookup table}
\usage{
load_region_lookup()
}
\value{
A data frame containing columns \code{Code} (SUBNATIONAL1_CODE) and \code{Area} (region name).
}
\description{
This function loads the region lookup table, which maps SUBNATIONAL1_CODE values to region "common" names.
This function mostly exists to be part of the pfw_region function.
}
