% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfw_example.R
\docType{data}
\name{pfw_example}
\alias{pfw_example}
\title{Example Project FeederWatch Dataset}
\format{
A data frame with 556,814 rows and 24 columns.
}
\source{
Created using \code{pfw_download()} and \code{pfw_import()} in data-raw/pfw_example.R
}
\usage{
pfw_example
}
\description{
A sample dataset for demonstration and testing purposes.
This dataset includes data from 2021 - May 2024 from
Washington and Oregon.
}
\examples{
# Load the example data into the environment
data(pfw_example)

# Assign the example dataset
testing_data <- pfw_example

}
\keyword{datasets}
