% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfw_import.R
\name{pfw_import}
\alias{pfw_import}
\title{Import Project FeederWatch Data}
\usage{
pfw_import(folder = "data-raw/", filter = FALSE, ...)
}
\arguments{
\item{folder}{The folder where PFW data is stored. Default is "data-raw/".}

\item{filter}{Logical. If TRUE, applies filters using pfw_filter(). Default is FALSE.}

\item{...}{Additional arguments passed to pfw_filter() for filtering (e.g., region, species, year).}
}
\value{
A combined and optionally filtered data.table containing all PFW data.
}
\description{
This function reads all .csv files downloaded from the PFW website, either from the default "data-raw/" folder
or from a user-specified folder. Optionally, it can apply filters like region, species, year, etc.
}
