% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfw_region.R
\name{pfw_region}
\alias{pfw_region}
\title{Filter Project FeederWatch Data by Region}
\usage{
pfw_region(data, regions)
}
\arguments{
\item{data}{The active PFW dataset (a data.table or data.frame) in your R project.}

\item{regions}{A character vector of regions (e.g., "Washington", "United States").}
}
\value{
A filtered dataset containing only the selected regions.
}
\description{
This function filters PFW data to include only specified states, provinces, or countries.
}
