% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfw_rollup.R
\name{pfw_rollup}
\alias{pfw_rollup}
\title{Do Taxonomic Rollup on Project FeederWatch Data}
\usage{
pfw_rollup(data)
}
\arguments{
\item{data}{A PFW dataset with a SPECIES_CODE column.}
}
\value{
A cleaned dataset with only species-level codes and a rollup attribute.
}
\description{
This function removes spuhs, hybrids, and slashes and
"demotes" subspecies/subspecies intergrades to their parent species.
}
