% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfw_sitedata.R
\name{pfw_sitedata}
\alias{pfw_sitedata}
\title{Merge Site Metadata into Project FeederWatch Data}
\usage{
pfw_sitedata(data, path)
}
\arguments{
\item{data}{A Project FeederWatch dataset (data.frame or data.table).}

\item{path}{File path to a the site description data file from https://feederwatch.org/explore/raw-dataset-requests/.}
}
\value{
The original dataset with site metadata merged in.
}
\description{
This function joins habitat and site metadata into PFW observation data using the site description file.
}
