% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfw_example.R
\name{pfw_example}
\alias{pfw_example}
\title{Download and Load Example Project FeederWatch Dataset}
\usage{
pfw_example()
}
\value{
A dataset of example Project FeederWatch data.
}
\description{
Provides a sample dataset for demonstration and testing purposes.
This dataset includes data from 2021 - May 2024 from
Washington, Oregon, and California.
}
\examples{
# Install and load the example dataset
example_data <- pfw_example()

}
