% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfw_rollup.R
\name{pfw_rollup}
\alias{pfw_rollup}
\title{Do Taxonomic Rollup on Project FeederWatch Data}
\usage{
pfw_rollup(data)
}
\arguments{
\item{data}{A Project FeederWatch dataset.}
}
\value{
A cleaned dataset with only species-level codes and a rollup attribute.
}
\description{
This function removes spuhs, hybrids, and slashes and
"demotes" subspecies/subspecies intergrades to their parent species.
}
\examples{
# Download/load example dataset
data <- pfw_example()

# Apply taxonomic rollup to an active PFW dataset
rolled_data <- pfw_rollup(data)

}
