% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfw_zerofill.R
\name{pfw_zerofill}
\alias{pfw_zerofill}
\title{Zerofill Species not Detected in each Survey Instance for Analysis}
\usage{
pfw_zerofill(data)
}
\arguments{
\item{data}{A Project FeederWatch dataset, optionally filtered for species.}
}
\value{
A data.table with zerofilled values included for each species.
}
\description{
This function adds zeros for checklists where selected species were absent,
setting HOW_MANY = 0 for presence/absence-based analyses.
Note that zerofilling entire, unfiltered datasets from Project FeederWatch will take a long time!
}
