% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxonomy.R
\name{update_taxonomy}
\alias{update_taxonomy}
\title{Update the Project FeederWatch Species Translation Table}
\usage{
update_taxonomy()
}
\value{
A message confirming whether the update was successful.
}
\description{
This function downloads the latest species translation table from the Project FeederWatch website
and saves it to \verb{inst/extdata/SpeciesTranslationTable/}. If a previous version exists, the user will be asked
for confirmation before overwriting it. This ensures taxonomy can readily be kept up to date annually,
since it will only be manually updated on the PFW website otherwise.
}
